SOLUCIÓN del External Match sin funcionar.
Como el mapa de registros ordena y separa cada 4 B cada registro, pero
para el caso de los CAP (Capture Input) por HW hay solamente 3 canales
disponibles en vez de 4, cuando hay 4 registros CAP, generó la confusión
de crear una macro que indique 3 canales disponibles del CAP. El mapa
propio de registros quedó con 3 canales, cuando hay 4 canales, aunque el
Switch Matrix o SWM no soporte más de 3 canales en HW a la vez.
Entonces, cada vez que se asignaba al EMR, no se asignaba al registro
pensado, sinó a uno desplazado 4 B.

>> Solución:
* Cambiar la macro por un 4.
* Cambiar funciones de configuración de Switch Matrix para trabajar con
"__Canales_CAP - 1" de límite.
* Probar.

>> Resultados:
Funcionó y el LED verde integrado (PIO1.0) parpadea a 0,5 s. :D

+++ LPC845_code/Mars_Rover-LPC845/inc/Drivers/C-Timer/c_timer.h
@@ -55,8 +55,8 @@
     	// ## Variables ##
     	private:
 			typedef struct MAT_data_s {
-				uint8_t				port;
-				uint8_t				pin;
+				int16_t				port;
+				int16_t				pin;
 				uint32_t			period;
 				__IO uint32_t		* const EMRx;
 				__IO uint32_t		* const MCRx;
@@ -65,8 +65,8 @@
 			} MAT_data_t;
 
 			typedef struct CAP_data_s {
-				uint8_t				port;
-				uint8_t				pin;
+				int16_t				port;
+				int16_t				pin;
 				__IO uint32_t 		* const CCRmode;
 				__IO uint32_t 		* const CTCRedge;
 				__I  uint32_t 		* const CRx;
